Nova-Long-Horizon-Agentic-Ai/requirements.txt
google-genai>=1.0.0
ollama>=0.3.0
rich>=13.7.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
typer>=0.12.0
aiofiles>=24.0.0

# LangChain
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0
tavily-python>=0.3.0
